requirements.txt
Pillow==11.0.0
numpy==2.1.3
matplotlib==3.9.2a ja 